refactor(cli): split spec-validation and state-db logic into dedicated modules
Extract flow-spec validation helpers from cli/orchestrate/__init__.py into
new _spec.py sibling; __init__.py re-exports all names so import paths are
unchanged. Split cli/state.py (~893 LOC) into _state_db_import.py (pure
per-run helpers) and _state_db_ops.py (ls/stats/checkpoint/vacuum/prune/
doctor), with state.py reduced to ~165 LOC dispatch layer that re-exports
the full public surface. RUNS_ROOT-dependent scan loops (_import_runs,
_import_teams) remain in state.py for monkeypatch-compatible test isolation.
No behavior changes; 782 tests pass.
