+++ lionagi/cli/state.py
@@ -6,36 +6,51 @@ from __future__ import annotations
 
 import argparse
 import json
-import uuid
-from pathlib import Path
-from typing import Any
 
-from ._lifecycle import EXIT_CODE_BY_STATUS
+# Import pure helpers from split modules (public surface preserved).
 from ._runs import RUNS_ROOT
+from ._state_db_import import (
+    _EXIT_CODE_STATUS_MAP,
+    _STATUS_MAP,
+    _derive_import_status,
+    _derive_timestamps,
+    _import_one_run,
+    _msg_from_collection_entry,
+    _mtime_as_float,
+)
 
+# Re-export DB operations helpers (public surface preserved).
+from ._state_db_ops import (
+    _checkpoint,
+    _doctor,
+    _format_bytes,
+    _list_sessions,
+    _print_stats,
+    _prune,
+    _vacuum,
+)
 
-def _mtime_as_float(path: Path) -> float:
-    try:
-        return path.stat().st_mtime
-    except OSError:
-        import time
-
-        return time.time()
-
-
-def _msg_from_collection_entry(raw: dict[str, Any]) -> dict[str, Any]:
-    """Convert a branch-collection message dict to the DB insert shape."""
-    return {
-        "id": raw["id"],
-        "created_at": raw["created_at"],
-        "node_metadata": raw.get("metadata"),
-        "content": raw.get("content", {}),
-        "embedding": raw.get("embedding"),
-        "sender": raw.get("sender"),
-        "recipient": raw.get("recipient"),
-        "channel": raw.get("channel"),
-        "role": raw["role"],
-    }
+__all__ = [
+    # import helpers
+    "RUNS_ROOT",
+    "_mtime_as_float",
+    "_msg_from_collection_entry",
+    "_import_runs",
+    "_STATUS_MAP",
+    "_EXIT_CODE_STATUS_MAP",
+    "_derive_import_status",
+    "_derive_timestamps",
+    "_import_one_run",
+    "_import_teams",
+    # ops helpers
+    "_format_bytes",
+    "_list_sessions",
+    "_print_stats",
+    "_checkpoint",
+    "_vacuum",
+    "_prune",
+    "_doctor",
+]
 
 
 async def _import_runs() -> dict[str, int]:
@@ -89,227 +104,10 @@ async def _import_runs() -> dict[str, int]:
     return counts
 
 
-_STATUS_MAP = {
-    "running": "running",
-    "completed": "completed",
-    "failed": "failed",
-    "aborted": "aborted",
-    "timed_out": "timed_out",
-    "cancelled": "cancelled",
-    "canceled": "cancelled",
-    "success": "completed",
-    "error": "failed",
-    "timeout": "timed_out",
-}
-
-_EXIT_CODE_STATUS_MAP = {v: k for k, v in EXIT_CODE_BY_STATUS.items()}
-
-
-def _derive_import_status(manifest: dict[str, Any]) -> str:
-    """Derive session status from run.json fields."""
-    raw_status = manifest.get("status")
-    if raw_status is not None:
-        return _STATUS_MAP.get(str(raw_status).lower(), "completed")
-
-    exit_code = manifest.get("exit_code")
-    if exit_code is not None:
-        return _EXIT_CODE_STATUS_MAP.get(exit_code, "failed")
-
-    return "completed"
-
-
-def _derive_timestamps(
-    manifest: dict[str, Any],
-    run_dir: Path,
-) -> tuple[float, float]:
-    """Return (started_at, ended_at) as floats; falls back to fs timestamps."""
-    import time as _time
-
-    started_at = manifest.get("started_at")
-    ended_at = manifest.get("ended_at")
-
-    try:
-        stat = run_dir.stat()
-        fs_ctime = stat.st_birthtime if hasattr(stat, "st_birthtime") else stat.st_ctime
-        fs_mtime = stat.st_mtime
-    except OSError:
-        now = _time.time()
-        fs_ctime = now
-        fs_mtime = now
-
-    if started_at is None:
-        started_at = fs_ctime
-    if ended_at is None:
-        ended_at = fs_mtime
-
-    if isinstance(started_at, str):
-        import datetime
-
-        try:
-            started_at = datetime.datetime.fromisoformat(started_at).timestamp()
-        except ValueError:
-            started_at = fs_ctime
-    if isinstance(ended_at, str):
-        import datetime
-
-        try:
-            ended_at = datetime.datetime.fromisoformat(ended_at).timestamp()
-        except ValueError:
-            ended_at = fs_mtime
-
-    return float(started_at), float(ended_at)
-
-
-async def _import_one_run(
-    db: Any,
-    run_id: str,
-    run_dir: Path,
-    manifest: dict[str, Any],
-) -> tuple[int, int, int]:
-    created_at = _mtime_as_float(run_dir)
-    session_name = manifest.get("kind") or "agent"
-
-    status = _derive_import_status(manifest)
-    started_at, ended_at = _derive_timestamps(manifest, run_dir)
-
-    session_prog_id = str(uuid.uuid4())
-    await db.create_progression(session_prog_id)
-
-    raw_kind = (manifest.get("kind") or "").lower()
-    legacy_kind_map = {
-        "agent": "agent",
-        "play": "play",
-        "flow": "flow",
-        "fanout": "fanout",
-    }
-    invocation_kind = legacy_kind_map.get(raw_kind)
-
-    artifacts_path = manifest.get("artifact_root") or manifest.get("artifacts_path")
-    if artifacts_path is None:
-        candidate = run_dir / "artifacts"
-        if candidate.exists():
-            artifacts_path = str(candidate)
-
-    await db.create_session(
-        {
-            "id": run_id,
-            "created_at": created_at,
-            "node_metadata": None,
-            "name": session_name,
-            "user": None,
-            "progression_id": session_prog_id,
-            "first_msg_id": None,
-            "last_msg_id": None,
-            # ADR-0012 enriched provenance — written so imported rows are
-            # queryable by the same fields live runs use.
-            "invocation_kind": invocation_kind,
-            "playbook_name": manifest.get("playbook_name") or manifest.get("playbook"),
-            "agent_name": manifest.get("agent_name") or manifest.get("agent"),
-            "artifacts_path": artifacts_path,
-            "source_kind": "imported_fs",
-            "status": status,
-            "started_at": started_at,
-            "ended_at": ended_at,
-        }
-    )
-
-    branches_dir = run_dir / "branches"
-
-    branch_files: list[Path] = []
-    if branches_dir.exists():
-        branch_files = list(branches_dir.glob("*.json"))
-
-    total_branches = 0
-    total_messages = 0
-    session_msg_ids: list[str] = []
-
-    for branch_file in sorted(branch_files, key=lambda p: p.stat().st_mtime):
-        try:
-            branch_data = json.loads(branch_file.read_text())
-        except Exception as exc:
-            print(f"    [warn] {branch_file.name}: failed to read — {exc}")
-            continue
-
-        branch_id = branch_data.get("id") or branch_file.stem
-        branch_created_at = branch_data.get("created_at") or _mtime_as_float(branch_file)
-
-        messages_pile = branch_data.get("messages", {})
-        raw_collection: list[dict] = messages_pile.get("collections", [])
-        progression_info = messages_pile.get("progression", {})
-        order: list[str] = progression_info.get("order", [])
-
-        by_id: dict[str, dict] = {m["id"]: m for m in raw_collection if "id" in m}
-        if order:
-            ordered_msgs = [by_id[mid] for mid in order if mid in by_id]
-        else:
-            ordered_msgs = raw_collection
-
-        system_msg_id: str | None = None
-        for raw_msg in ordered_msgs:
-            if raw_msg.get("role") == "system":
-                system_msg_id = raw_msg["id"]
-                break
-
-        branch_msg_ids: list[str] = []
-        for raw_msg in ordered_msgs:
-            msg = _msg_from_collection_entry(raw_msg)
-            await db.insert_message(msg)
-            branch_msg_ids.append(msg["id"])
-            total_messages += 1
-
-        # Create branch progression with ordered message IDs.
-        branch_prog_id = str(uuid.uuid4())
-        await db.create_progression(branch_prog_id, branch_msg_ids)
-
-        manifest_branch_meta = {}
-        for mb in manifest.get("branches", []):
-            if mb.get("id") == branch_id:
-                manifest_branch_meta = mb
-                break
-
-        node_meta: dict[str, Any] = {}
-        provider = manifest_branch_meta.get("provider") or manifest.get("provider")
-        model = manifest_branch_meta.get("model") or manifest.get("model")
-        if provider:
-            node_meta["provider"] = provider
-        if model:
-            node_meta["model"] = model
-        branch_name = manifest_branch_meta.get("name") or manifest.get("kind")
-
-        await db.create_branch(
-            {
-                "id": branch_id,
-                "created_at": branch_created_at,
-                "node_metadata": node_meta or None,
-                "user": branch_data.get("user"),
-                "name": branch_name,
-                "session_id": run_id,
-                "progression_id": branch_prog_id,
-                "system_msg_id": system_msg_id,
-            }
-        )
-
-        session_msg_ids.extend(branch_msg_ids)
-        total_branches += 1
-
-    if session_msg_ids:
-        await db.db.execute(
-            "UPDATE progressions SET collection = ? WHERE id = ?",
-            (json.dumps(session_msg_ids), session_prog_id),
-        )
-        await db.db.commit()
-        await db.update_session(
-            run_id,
-            first_msg_id=session_msg_ids[0],
-            last_msg_id=session_msg_ids[-1],
-        )
-
-    print(f"  imported {run_id}: {total_branches} branch(es), {total_messages} message(s)")
-    return 1, total_branches, total_messages
-
-
 async def _import_teams() -> dict[str, int]:
     """Backfill ~/.lionagi/teams/*.json into the teams + team_messages tables."""
+    import uuid
+
     from lionagi.state.db import StateDB
 
     teams_dir = (RUNS_ROOT.parent / "teams").resolve()
@@ -400,264 +198,6 @@ async def _import_teams() -> dict[str, int]:
         await db.db.commit()
 
     return counts
-
-
-async def _list_sessions(*, limit: int = 50, status: str | None = None) -> None:
-    import time
-
-    from lionagi.state.db import StateDB
-
-    async with StateDB() as db:
-        if status:
-            cur = await db.db.execute(
-                "SELECT id, name, status, updated_at FROM sessions "
-                "WHERE status = ? ORDER BY updated_at DESC LIMIT ?",
-                (status, limit),
-            )
-        else:
-            cur = await db.db.execute(
-                "SELECT id, name, status, updated_at FROM sessions "
-                "ORDER BY updated_at DESC LIMIT ?",
-                (limit,),
-            )
-        rows = await cur.fetchall()
-
-        if not rows:
-            print("(no sessions in state.db)")
-            return
-
-        header = (
-            f"{'ID':<36}  {'NAME':<16}  {'STATUS':<10}  "
-            f"{'BRANCHES':>8}  {'MESSAGES':>8}  {'UPDATED':<20}"
-        )
-        print(header)
-        print("-" * len(header))
-        for row in rows:
-            sid = row["id"]
-            name = (row["name"] or "")[:16]
-            sstat = (row["status"] or "")[:10]
-            updated = row["updated_at"]
-            updated_str = (
-                time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(updated)) if updated else ""
-            )
-
-            branch_cur = await db.db.execute(
-                "SELECT COUNT(*) AS n FROM branches WHERE session_id = ?", (sid,)
-            )
-            bc = (await branch_cur.fetchone())["n"]
-
-            prog_cur = await db.db.execute(
-                "SELECT progression_id FROM sessions WHERE id = ?", (sid,)
-            )
-            prog_row = await prog_cur.fetchone()
-            msg_count = 0
-            if prog_row and prog_row["progression_id"]:
-                prog_data = await db.get_progression(prog_row["progression_id"])
-                msg_count = len(prog_data)
-
-            print(f"{sid:<36}  {name:<16}  {sstat:<10}  {bc:>8}  {msg_count:>8}  {updated_str:<20}")
-
-
-async def _print_stats() -> None:
-    from lionagi.state.db import DEFAULT_DB_PATH, StateDB
-
-    db_path = DEFAULT_DB_PATH
-    db_size = db_path.stat().st_size if db_path.exists() else 0
-    wal_path = db_path.with_name(db_path.name + "-wal")
-    wal_size = wal_path.stat().st_size if wal_path.exists() else 0
-
-    print(f"state.db path:   {db_path}")
-    print(f"state.db size:   {_format_bytes(db_size)}")
-    print(f"state.db-wal:    {_format_bytes(wal_size)}")
-    print()
-
-    if not db_path.exists():
-        print("(no state.db yet — first run will create it)")
-        return
-
-    async with StateDB() as db:
-        print("Row counts:")
-        for table in (
-            "messages",
-            "progressions",
-            "sessions",
-            "branches",
-            "definitions",
-            "shows",
-            "plays",
-        ):
-            cur = await db.db.execute(
-                f"SELECT COUNT(*) AS n FROM {table}"  # noqa: S608
-            )
-            row = await cur.fetchone()
-            print(f"  {table:<14} {row['n']:>10}")
-        print()
-
-        cur = await db.db.execute(
-            "SELECT COALESCE(status, '(null)') AS s, COUNT(*) AS n "
-            "FROM sessions GROUP BY status ORDER BY n DESC"
-        )
-        print("Sessions by status:")
-        for row in await cur.fetchall():
-            print(f"  {row['s']:<14} {row['n']:>10}")
-        print()
-
-        print("PRAGMAs:")
-        for pragma in (
-            "journal_mode",
-            "wal_autocheckpoint",
-            "busy_timeout",
-            "synchronous",
-            "foreign_keys",
-        ):
-            cur = await db.db.execute(f"PRAGMA {pragma}")
-            row = await cur.fetchone()
-            val = row[0] if row else "?"
-            print(f"  {pragma:<22} {val}")
-
-
-async def _checkpoint(mode: str) -> str:
-    from lionagi.state.db import StateDB
-
-    async with StateDB() as db:
-        cur = await db.db.execute(f"PRAGMA wal_checkpoint({mode})")
-        row = await cur.fetchone()
-        if not row:
-            return "(no result)"
-        return f"busy={row[0]}, log_pages={row[1]}, checkpointed={row[2]}"
-
-
-async def _vacuum() -> None:
-    from lionagi.state.db import StateDB
-
-    async with StateDB() as db:
-        await db.db.execute("VACUUM")
-        await db.db.commit()
-
-
-async def _prune(
-    *,
-    keep_days: int,
-    keep_n: int,
-    dry_run: bool,
-) -> dict[str, int]:
-    import time as _time
-
-    from lionagi.state.db import StateDB
-
-    cutoff = _time.time() - (keep_days * 86400)
-
-    async with StateDB() as db:
-        cur = await db.db.execute(
-            """SELECT id FROM sessions
-               WHERE id NOT IN (
-                 SELECT id FROM sessions
-                 ORDER BY updated_at DESC LIMIT ?
-               )
-               AND (updated_at < ? OR updated_at IS NULL)""",
-            (keep_n, cutoff),
-        )
-        rows = await cur.fetchall()
-        victim_ids = [r["id"] for r in rows]
-
-        if not victim_ids:
-            return {"sessions": 0, "branches": 0, "messages": 0}
-
-        placeholders = ",".join("?" * len(victim_ids))
-        cur = await db.db.execute(
-            f"SELECT COUNT(*) AS n FROM branches "  # noqa: S608
-            f"WHERE session_id IN ({placeholders})",
-            victim_ids,
-        )
-        branch_count = (await cur.fetchone())["n"]
-
-        cur = await db.db.execute("SELECT COUNT(*) AS n FROM messages")
-        msgs_before = (await cur.fetchone())["n"]
-
-        if dry_run:
-            return {
-                "sessions": len(victim_ids),
-                "branches": branch_count,
-                "messages": 0,  # can't preview without doing the delete
-            }
-
-        await db.db.execute(
-            f"DELETE FROM sessions WHERE id IN ({placeholders})",  # noqa: S608
-            victim_ids,
-        )
-        await db.db.commit()
-
-        await db.db.execute(
-            """DELETE FROM messages
-               WHERE id NOT IN (
-                 SELECT value FROM progressions, json_each(progressions.collection)
-               )"""
-        )
-        await db.db.commit()
-
-        cur = await db.db.execute("SELECT COUNT(*) AS n FROM messages")
-        msgs_after = (await cur.fetchone())["n"]
-
-        return {
-            "sessions": len(victim_ids),
-            "branches": branch_count,
-            "messages": msgs_before - msgs_after,
-        }
-
-
-async def _doctor(
-    *,
-    stale_hours: int,
-    dry_run: bool,
-    new_status: str = "aborted",
-) -> dict[str, int]:
-    """Sweep sessions stuck at status='running' older than stale_hours."""
-    import time as _time
-
-    from lionagi.state.db import StateDB
-
-    cutoff = _time.time() - (stale_hours * 3600)
-
-    async with StateDB() as db:
-        cur = await db.db.execute("SELECT id, started_at FROM sessions WHERE status = 'running'")
-        rows = await cur.fetchall()
-        total = len(rows)
-        victims: list[str] = []
-        skipped = 0
-        for row in rows:
-            started = row["started_at"]
-            if started is None or started < cutoff:
-                victims.append(row["id"])
-            else:
-                skipped += 1
-
-        swept_count = 0
-        if dry_run:
-            swept_count = len(victims)
-        elif victims:
-            # Re-assert status='running' in the UPDATE to avoid race with
-            # sessions that finish between select and update.
-            placeholders = ",".join("?" * len(victims))
-            params = [new_status, _time.time(), cutoff, *victims]
-            cur = await db.db.execute(
-                f"UPDATE sessions SET status = ?, ended_at = ? "  # noqa: S608
-                f"WHERE status = 'running' "
-                f"  AND (started_at IS NULL OR started_at < ?) "
-                f"  AND id IN ({placeholders})",
-                params,
-            )
-            swept_count = cur.rowcount or 0
-            await db.db.commit()
-
-        return {"running": total, "swept": swept_count, "skipped": skipped}
-
-
-def _format_bytes(n: int) -> str:
-    for unit in ("B", "KiB", "MiB", "GiB"):
-        if n < 1024:
-            return f"{n:.1f} {unit}"
-        n /= 1024
-    return f"{n:.1f} TiB"
 
 
 def add_state_subparser(subparsers: argparse._SubParsersAction) -> None:
